model: split artifact and residency namespaces
