+++ crates/izwi-core/src/residency/mod.rs
