+++ crates/izwi-core/src/runtime_models/mod.rs
@@ -3,4 +3,4 @@
 //! Public callers should prefer this namespace over the internal `models`
 //! implementation module.
 
-pub use crate::models::{architectures, registry, shared, ModelRegistry};
+pub use crate::models::{architectures, registry, shared, LoadedModelRegistry, ModelRegistry};
